requirements.txt
discord-webhook
pyyaml